fix: recognize test-file conventions across 7 languages
Dogfooding entropyx against a real Go/TS repo surfaced a bug: T_c was
zero for all 434 files even though the repo has dozens of *_test.go
and *.test.ts files. Root cause: is_test_path only recognized
Rust-layout conventions (tests/, tests.rs, _test.rs suffixes).

Extends the heuristic to cover all seven supported languages:
  - Go:     *_test.go
  - JS/TS:  *.test.{js,jsx,ts,tsx,mjs,cjs}, *.spec.{...}, __tests__/
  - Python: test_*.py, *_test.py, tests.py, test.py
  - Ruby:   *_spec.rb, *_test.rb, spec/
  - Java:   *Test.java, *Tests.java, *Spec.java, test/
  - C++:    *_test.{cc,cpp,cxx,hpp,h}, *_tests.{...}
  - Rust:   (unchanged — tests/, tests.rs, _test{s,}.rs, _spec.rs)

Also sweeps three clippy warnings flagged during the stabilization
pass: map_or → is_some_and (two sites), needless `&` borrow on a
Copy-compatible argument, and an unnecessary u32→u32 cast.

Re-scanning RoomIQ after the fix shows T_c=1.0 on 62 test files
(previously 14) and on 296 co-tested source files (previously 0).

+++ crates/entropyx-git/src/repo.rs
@@ -303,7 +303,7 @@ fn iter_diff(from_tree: &gix::Tree<'_>, to_tree: &gix::Tree<'_>) -> Result<Vec<F
                 let from = String::from_utf8_lossy(source_location).into_owned();
                 let similarity = diff
                     .map(|d| {
-                        let total = (d.removals + d.insertions) as u32;
+                        let total = d.removals + d.insertions;
                         if total == 0 {
                             100
                         } else {
